Done with the day 9 challenge
In day 9, we were given a layout and asked to do it with flex, and hence did it.

P.S. KP used margin instead of gap in flex items because at the time of making the video, many browsers did not support it. But since browsers support it widely now, I used gap.